Organized MDeditor component
